+++ Changelog.md
@@ -23,6 +23,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - FIO-7786: Fixing Datagrid issue in Settings JSON
  - FIO-4905: File upload (multi) - aborting upload always cancels the last one in the list
  - FIO-7642: fixed issues where calculated value with allow override is not recalculated after form/component/row values are reset
+ - FIO-7632: Fixes an issue where HTML tags are added to the HTML5 Select metadata
+ - FIO-4871: fixed calculated value issues
+ - FIO 7603: fixed Edit Grid With Empty Rows Not Submitting Form
+ - FIO-7445: fixed an issue where the interpolated data does not show up on PDF
+ - FIO-7421: Adds ReCaptcha error messages to the translations config
+ - FIO-7804: Added PKCE method for OIDC
 
 ## 5.0.0-rc.37
 ### Fixed